requirements.txt
playwright==1.40.0
gspread==5.12.0
google-auth==2.23.4
